pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[project]
name = "microgrowagents"
description = "Agent-based system for microbial growth predictions"
authors = [
  {name = "marcin p. joachimiak", email = "4625870+realmarcin@users.noreply.github.com"},
]
license = "BSD-3-Clause"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dynamic = ["version"]

dependencies = [
  "typer >= 0.9.0",
  "linkml-runtime >=1.9.4",
  "linkml >= 1.9.3",
  "click >= 8.0.0",
  "kghub-downloader >= 0.3.0",
  "duckdb >= 0.9.0",
  "pandas >= 2.0.0",
  "requests >= 2.31.0",
  "requests-cache >= 1.1.0",
  "pyyaml >= 6.0",
  "numpy >= 1.24.0",
  "scipy >= 1.10.0",
  "matplotlib >= 3.5.0",
  "seaborn >= 0.12.0",
  "tqdm >= 4.65.0",
  "beautifulsoup4 >= 4.12.0",
  "lxml >= 4.9.0",
  "fuzzywuzzy >= 0.18.0",
  "python-Levenshtein >= 0.21.0",
  "grape >= 0.1.0", # GRAPE: Graph Representation leArning (Rust-backed, optimized for large biological networks)
  "pypdf2>=3.0.1",
  "markitdown>=0.1.4",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "ruff>=0.4.8",
    "mkdocs-material>=8.2.8",
    "mkdocs-mermaid2-plugin>=1.1.1",
    "jupyter>=1.0.0",
    "mknotebooks>= 0.8.0",
]

[project.scripts]
MicroGrowAgents = "microgrowagents.cli:main"
microgrowagents = "microgrowagents.cli:main"


# See https://hatch.pypa.io/latest/config/build/#file-selection for how to
# explicitly include files other than default into the build distributions.

[tool.hatch.version]
source = "uv-dynamic-versioning"

# Ref.: https://github.com/ninoseki/uv-dynamic-versioning/
[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
fallback-version = "0.0.0"

# Ref.: https://docs.pytest.org/en/stable/reference/reference.html#configuration-options
[tool.pytest.ini_options]
testpaths = ["tests"]

# Ref.: https://github.com/codespell-project/codespell
[tool.codespell]
skip = [
  "LICENSE",
  "pyproject.toml",
  "uv.lock",
  "project/*",
  "src/microgrowagents/datamodel/microgrowagents_pydantic.py",
  "src/microgrowagents/datamodel/microgrowagents.py",
]

# Reminder: words have to be lowercased for the ignore-words-list
ignore-words-list = "linke"
quiet-level = 3

# Ref.: https://github.com/crate-ci/typos (spell checker)
[tool.typos.default.extend-words]
linke = "linke"
[tool.typos.files]
extend-exclude = [
  "LICENSE",
  "uv.lock",
  "pyproject.toml",
  "project/*",
  "src/microgrowagents/datamodel/microgrowagents_pydantic.py",
  "src/microgrowagents/datamodel/microgrowagents.py",
]
